Generate Content-Type headers for static file serving
This uses the `magic-mime` library to detect the filename and
generate a sensible MIME type from the `mime.types` database.
It is particularly important to generate these for Internet Explorer
support, since IE10 seems to ignore `.css` files that do not send
an appropriate MIME header.  With this patch, IE10 rendering works
great using an Opium rendering stack.